Trigger a real measurement.
Added a new type of policy, on replying the distance, the kernel should only reply u64 type.

Removed pseudo timer, previous we used this timer to mock distance measurement result.

In pin configuration request, we will register ISR to handle the interrupt.

Before we start the measurement, we saved the context information to the per-device structure, so the particular device can respond to the corresponding sender (socket).